+++ packages/terra-divider/src/terra-dev-site/test/divider/DefaultDivider.test.jsx
@@ -2,4 +2,10 @@ import React from 'react';
 
 import Divider from '../../../Divider';
 
-export default () => <Divider id="divider" />;
+export default () => (
+  <div>
+    <div style={{ height: '2rem', background: '#40E0D0' }} />
+    <Divider id="divider" />
+    <div style={{ height: '2rem', background: '#40E0D0' }} />
+  </div>
+);