+++ packages/terra-divider/src/terra-dev-site/test/divider/DividerWithWrappingText.test.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Divider from '../../../Divider';
+
+const content = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut enim ante, porta sit amet sem vitae, pellentesque pharetra ex. Etiam odio purus, maximus eget mauris in, pulvinar euismod neque. Curabitur eu vulputate leo. Etiam tincidunt lectus ut metus interdum, sit amet porttitor leo ornare. Sed tincidunt rutrum odio, dignissim rhoncus nulla finibus et. Mauris mollis posuere dolor dignissim vulputate. Sed accumsan dignissim mi id pulvinar.';
+
+export default () => (
+  <div>
+    <div style={{ height: '2rem', background: '#40E0D0' }} />
+    <Divider id="divider" text={content} />
+    <div style={{ height: '2rem', background: '#40E0D0' }} />
+  </div>
+);