+++ packages/terra-form-input/src/InputField.jsx
@@ -83,6 +83,10 @@ const propTypes = {
    */
   showOptional: PropTypes.bool,
   /**
+   * Specifies the type of input element to display.
+   */
+  type: PropTypes.string,
+  /**
    * The value of the input field. Use this to create a controlled input.
    */
   value: PropTypes.oneOfType([
@@ -108,6 +112,7 @@ const defaultProps = {
   refCallback: undefined,
   required: false,
   showOptional: false,
+  type: undefined,
   value: undefined,
 };
 
@@ -131,6 +136,7 @@ const InputField = (props) => {
     refCallback,
     required,
     showOptional,
+    type,
     value,
     ...customProps
   } = props;
@@ -149,6 +155,7 @@ const InputField = (props) => {
     }
   }
 
+  const inputType = type || inputAttrs.type;
   return (
     <Field
       label={label}
@@ -170,6 +177,7 @@ const InputField = (props) => {
         {...inputAttrs}
         disabled={inputAttrs.disabled || disabled}
         id={inputId}
+        type={inputType}
         onChange={onChange}
         value={value}
         defaultValue={defaultValue}