+++ packages/terra-base/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+5.3.0 - (April 16, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 5.2.0 - (April 9, 2019)
 ------------------
 ### Changed