+++ packages/terra-form-input/CHANGELOG.md
@@ -6,6 +6,15 @@ Unreleased
 ### Changed
 * Added 'disabled' prop to InputField component
 
+2.11.0 - (April 16, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Added type as first class prop to Form-Input-Field
+* Added example of Number Input on for Form-Input
+
 2.10.0 - (April 9, 2019)
 ------------------
 ### Changed