+++ packages/terra-form-select/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Added 'disabled' prop to FormSelectField component
 
+5.14.0 - (April 16, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 5.13.0 - (April 9, 2019)
 ------------------
 ### Fixed