+++ packages/terra-form-textarea/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Added 'disabled' prop to TextAreaField component
 
+3.11.0 - (April 16, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.10.0 - (April 9, 2019)
 ------------------
 ### Changed