+++ packages/terra-list/CHANGELOG.md
@@ -4,13 +4,18 @@ Changelog
 Unreleased
 ----------
 
+4.5.0 - (April 16, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 ### Added
 * Added an interaction media query to apply `hover` styles over a list-item only on non-touch devices.
 
 4.4.0 - (April 9, 2019)
 ------------------
 ### Changed
-* Update documentation to better cover sections and headers
+* Update documentation to better cover sections and headers and for guides to be more interactive
 
 4.3.0 - (March 26, 2019)
 ------------------