+++ packages/terra-visually-hidden-text/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+2.6.0 - (April 16, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.5.0 - (March 26, 2019)
 ------------------
 ### Changed